moved ListOwner to model module
